+++ projects/cloud/app/frontend/components/NoPageFound.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NoPageFound = () => {
+  return <p>Page not found</p>;
+};
+
+export default NoPageFound;
